Application/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0+cpu
torchvision==0.21.0+cpu
transformers==4.38.2
flask==3.0.2
pillow
